Update button.md
Under the parameters section, it would be beneficial to actually list some examples of how to set each and every one of them. For example the parameter -Size just says it is of type String and the Description says 'the size of the button'. This is confusing for newcomers who do not understand what options they have for that string; I have been doing some digging for some time on the forum, and have yet to find clarification.

Perhaps it would be better to drop Size altogether in lieu of the Style hashtable. If this route is taken, however, there should be a concrete example of use as I added above