Update grid-layout.md
As a relatively new user, I think moving the above information from the github page into the main documentation would be extremely helpful. This saves from having to go to an external source for useful information on a function. Also, if possible it would be great to provide the ability to copy the JSON to an external file and have the -Layout parameter import this. This may make portability of scripts more easy than a long (in my case, if wrapped, 88 lines) string of JSON text in the script itself.